% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{D4File}
\alias{D4File}
\alias{$.D4File}
\title{D4File class}
\format{
An object of class \code{environment} of length 5.
}
\usage{
D4File
}
\description{
D4File class
}
\examples{
d4 <- D4File$new(path)
}
\keyword{datasets}
