% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{D4Source}
\alias{D4Source}
\alias{$.D4Source}
\title{D4Source class}
\format{
An object of class \code{environment} of length 11.
}
\usage{
D4Source
}
\description{
This is the primary way of interacting with D4 sources.
}
\examples{
d4 <- D4File$new(path)
chroms <- d4$get_chroms()
}
\keyword{datasets}
