% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{D4SourceEnv}
\alias{D4SourceEnv}
\alias{$.D4SourceEnv}
\title{D4SourceEnv environment}
\format{
An object of class \code{environment} of length 10.
}
\usage{
D4SourceEnv
}
\description{
Environment with bindings to functions for opening a connection to a D4 file and querying the file.
Generated by \code{rextendr} to expose bindings into Rust functions. Most users should not use this enviornment directly,
but should use \code{D4Source} and associated methods instead. In particular, documentation is provided along with \code{ D4Source} and its methods.
}
\keyword{datasets}
