% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{get_chroms.D4Source}
\alias{get_chroms.D4Source}
\title{Returns the list of chromosomes in the source D4 file}
\usage{
\method{get_chroms}{D4Source}(d4source)
}
\value{
Nested list; each element represents a chromosome and stores the chromosome name and length
}
\description{
Returns the list of chromosomes in the source D4 file
}
\examples{
d4source <- D4Source(system.file("extdata", "example.d4", package="rd4"))
get_chroms(d4source)

}
