% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{get_chroms}
\alias{get_chroms}
\title{Returns the list of chromosomes in the source D4 file}
\usage{
get_chroms(d4source)
}
\arguments{
\item{d4source}{\code{D4Source} object}
}
\value{
Nested list; each element represents a chromosome and stores the
chromosome name and length
}
\description{
Returns the list of chromosomes in the source D4 file
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
get_chroms(d4source)

}
