% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{get_source.D4Source}
\alias{get_source.D4Source}
\title{Returns the path to the source D4 file}
\usage{
\method{get_source}{D4Source}(d4source)
}
\value{
The path to the source D4 file
}
\description{
Returns the path to the source D4 file
}
\examples{
d4source <- D4Source(system.file("extdata", "example.d4", package="rd4"))
get_source(d4source)

}
