% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{get_source}
\alias{get_source}
\title{Get the path to a source D4 file}
\usage{
get_source(d4source)
}
\arguments{
\item{d4source}{\code{D4Source} object}
}
\value{
The path to the source D4 file
}
\description{
Get the path to a source D4 file
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
get_source(d4source)

}
