% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{get_tracks}
\alias{get_tracks}
\title{Get the names of tracks in a source D4 file}
\usage{
get_tracks(d4source)
}
\arguments{
\item{d4source}{\code{D4Source} object}
}
\value{
Vector of track names
}
\description{
Get the names of tracks in a source D4 file
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
get_tracks(d4source)

}
