% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Histogram}
\alias{Histogram}
\alias{$.Histogram}
\title{The \code{\link{Histogram}} API.}
\format{
An object of class \code{environment} of length 8.
}
\usage{
Histogram
}
\description{
The \code{\link{Histogram}} API.
}
\keyword{datasets}
