% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{HistogramEnv}
\alias{HistogramEnv}
\alias{$.HistogramEnv}
\title{HistogramEnv Environment}
\format{
An object of class \code{environment} of length 8.
}
\usage{
HistogramEnv
}
\description{
Environment with bindings to functions for accessing the attributes of a histogram. Generated by \code{rextendr} to
expose bindings into Rust functions. Most users should not use this environment directly, but should instead call
S3 methods on a \code{D4Source} object to return features of the distribution.
}
\examples{
path <- system.file("extdata", "example.d4", package="rd4")
d4source_env <- D4SourceEnv$new(path)
histogram_env <- d4source_env$histogram("chr1", 1000, 2000, track = NA, min = 0, max = 100)
histogram_env$total_count()

}
\keyword{datasets}
