% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{mean.D4Source}
\alias{mean.D4Source}
\title{Get the mean data value over the given region}
\usage{
mean.D4Source(x, ..., chr, start, end, track = NA)
}
\arguments{
\item{x}{\code{D4Source} object}

\item{...}{Not used for values, forces later arguments to bind by name}

\item{chr}{Chromosome/contig name}

\item{start}{1-based start position}

\item{end}{1-based inclusive end position}

\item{track}{Track name. If omitted or NA, queries the first track.}
}
\value{
Mean data value
}
\description{
Get the mean data value over the given region
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
mean(d4source, chr = 'chr1', start = 1000, end = 2000)

}
