% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{median.D4Source}
\alias{median.D4Source}
\title{Get the median data value over the given region}
\usage{
median.D4Source(x, na.rm = FALSE, ..., chr, start, end, track = NA)
}
\arguments{
\item{x}{\code{D4Source} object}

\item{na.rm}{Ignored; included in signature only to enable preservation of
\code{stats::median} in the namespace while implementing median for D4Source}

\item{...}{Not used for values, forces later arguments to bind by name}

\item{chr}{Chromosome/contig name}

\item{start}{1-based start position}

\item{end}{1-based inclusive end position}

\item{track}{Track name. If omitted or NA, queries the first track.}
}
\value{
Median data value
}
\description{
Get the median data value over the given region
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
median(d4source, chr = 'chr1', start = 1000, end = 2000)

}
