% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{percentile.D4Source}
\alias{percentile.D4Source}
\title{Returns the data value at the given percentile for the given region}
\usage{
\method{percentile}{D4Source}(d4source, chr, start, end, percentile, track = NA)
}
\arguments{
\item{d4source}{\code{D4Source} object}

\item{chr}{Chromosome/contig name}

\item{start}{0-based start position}

\item{end}{0-based exclusive end position}

\item{percentile}{Percentile between 0 and 100. Example: percentile 0 is the minimum data value, 50 is the median and
100 is the maximum.}

\item{track}{Track name. If omitted or NA, queries the first track.}
}
\value{
Data value at the given percentile
}
\description{
Returns the data value at the given percentile for the given region
}
\examples{
d4source <- D4Source(system.file("extdata", "example.d4", package="rd4"))
percentile(d4source, "chr1", 1000, 2000, 50)

}
