% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{print.D4Source}
\alias{print.D4Source}
\title{Print information about a \code{D4Source} object}
\usage{
\method{print}{D4Source}(x, ...)
}
\arguments{
\item{x}{\code{D4Source} object}

\item{...}{further arguments passed to or from other methods}
}
\value{
(Invisibly) the string that was printed
}
\description{
Print information about a \code{D4Source} object
}
