% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{query.D4Source}
\alias{query.D4Source}
\title{Extracts the given region from the D4 file}
\usage{
\method{query}{D4Source}(d4source, chr, start, end, track = NA)
}
\arguments{
\item{d4source}{\code{D4Source} object}

\item{chr}{Chromosome/contig name}

\item{start}{0-based start position}

\item{end}{0-based exclusive end position}

\item{track}{Track name. If omitted or NA, queries the first track.}
}
\value{
Query result as a list
}
\description{
Extracts the given region from the D4 file
}
\examples{
d4source <- D4Source(system.file("extdata", "example.d4", package="rd4"))
query_result <- query(d4source, "chr1", 1000, 2000)

}
