% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{QueryEnv}
\alias{QueryEnv}
\alias{$.QueryEnv}
\title{QueryEnv environment}
\format{
An object of class \code{environment} of length 3.
}
\usage{
QueryEnv
}
\description{
Environment with bindings to functions for accessing the parameters of a query. Generated by \code{rextendr} to expose
bindings into Rust functions. Most users should not use this environment directly, but should instead access the
$query attribute of a query result.
}
\keyword{datasets}
