% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{QueryResult}
\alias{QueryResult}
\alias{$.QueryResult}
\title{The object returned after querying a \code{\link{D4Source}}}
\format{
An object of class \code{environment} of length 5.
}
\usage{
QueryResult
}
\description{
The object returned after querying a \code{\link{D4Source}}
}
\keyword{datasets}
