% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{QueryResultEnv}
\alias{QueryResultEnv}
\alias{$.QueryResultEnv}
\title{QueryResultEnv Environment}
\format{
An object of class \code{environment} of length 5.
}
\usage{
QueryResultEnv
}
\description{
Environment with bindings to functions for extracting attributes from a query result. Generated by \code{rextendr} to
expose bindings into Rust functions. Most users should not use this environment directly, but should instead
access attributes of the list returned by \code{query()}.
}
\keyword{datasets}
