% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{resample}
\alias{resample}
\title{Generic method to resample values over a genomic data object}
\usage{
resample(
  x,
  chr,
  start,
  end,
  method,
  ...,
  bin_size,
  allow_bin_size_adjustment,
  track
)
}
\arguments{
\item{x}{Input object}

\item{chr}{Chromosome/contig name}

\item{start}{0-based start position}

\item{end}{0-based exclusive end position}

\item{method}{Summary statistic for the data in each bin}

\item{...}{Not used for values, forces later arguments to bind by name}

\item{bin_size}{Bin size}

\item{allow_bin_size_adjustment}{Allow bin size to be optimized}

\item{track}{Track name}
}
\value{
Resampled data
}
\description{
Generic method to resample values over a genomic data object
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
resample(d4source, 'chr1', 0, 50000000, 'mean')

}
