% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{update_median.D4Source}
\alias{update_median.D4Source}
\title{For a set of genomic ranges in a \code{GRanges} object, adds the median value
over the region as a metadata attribute for each genomic range, returning a
new \code{GRanges} object with updated metadata.}
\usage{
\method{update_median}{D4Source}(x, granges, ..., track = NA)
}
\arguments{
\item{x}{\code{D4Source} object}

\item{granges}{\code{GRanges} object containing genomic ranges of interest}

\item{...}{Not used for values, forces later arguments to bind by name}

\item{track}{Track name. If omitted or NA, queries the first track.}
}
\value{
New version of \code{granges} with region median added as the \code{'median'}
metadata attribute for each genomic range. If a track is specified, the
track name will be included in the metadata column name.
}
\description{
For a set of genomic ranges in a \code{GRanges} object, adds the median value
over the region as a metadata attribute for each genomic range, returning a
new \code{GRanges} object with updated metadata.
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
granges <- GRanges(seqnames = c('chr1'), ranges = IRanges(1000:2000), 
                   strand = c('+'))
granges_with_median <- update_median(d4source, granges)

}
