% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{update_percentile.D4Source}
\alias{update_percentile.D4Source}
\title{For a set of genomic ranges in a \code{GRanges} object, adds the percentile value over the region as a metadata attribute
for each genomic range, returning a new \code{GRanges} object with updated metadata.}
\usage{
\method{update_percentile}{D4Source}(d4source, granges, percentile, ..., track = NA)
}
\arguments{
\item{d4source}{\code{D4Source} object}

\item{granges}{\code{GRanges} object containing genomic ranges of interest}

\item{percentile}{Percentile between 0 and 100. Example: percentile 0 is the minimum data value, 50 is the median and
100 is the maximum.}

\item{track}{Track name. If omitted or NA, queries the first track.}
}
\value{
New version of \code{granges} with region percentile added as the \code{"percentile_<percentile>"} metadata attribute
for each genomic range. The percentile requested (between 0 and 100) is included in the metadata column name. If a
track is specified, the track name will also be included in the column name.
}
\description{
For a set of genomic ranges in a \code{GRanges} object, adds the percentile value over the region as a metadata attribute
for each genomic range, returning a new \code{GRanges} object with updated metadata.
}
\examples{
d4source <- D4Source(system.file("extdata", "example.d4", package="rd4"))
granges <- GRanges(seqnames=c("chr1"), ranges=IRanges(1000:2000), strand=c("+"))
granges_with_percentile <- update_percentile(d4source, granges, 50)

}
