% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{update_percentile}
\alias{update_percentile}
\title{Generic function to update percentile metadata in a GRanges object}
\usage{
update_percentile(x, granges, percentile, ..., track)
}
\arguments{
\item{x}{Input object}

\item{granges}{GRanges object}

\item{percentile}{Percentile between 0 and 100}

\item{...}{Not used for values, forces later arguments to bind by name}

\item{track}{Track name}
}
\value{
New version of \code{granges} with metadata updated
}
\description{
Generic function to update percentile metadata in a GRanges object
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
granges <- GRanges(seqnames = c('chr1'), ranges = IRanges(1000:2000), 
                   strand = c('+'))
granges_with_percentile <- update_percentile(d4source, granges, 50)

}
