% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{update_with_mean}
\alias{update_with_mean}
\title{Add mean D4 data values to a \code{GRanges} object}
\usage{
update_with_mean(granges, d4source, ..., track = NA, colname = NA)
}
\arguments{
\item{granges}{\code{GRanges} object containing genomic ranges of interest}

\item{d4source}{\code{D4Source} object}

\item{...}{Not used for values, forces later arguments to bind by name}

\item{track}{Track name. If omitted or NA, queries the first track.}

\item{colname}{Name of new metadata column to add. If omitted: the name of
the new metadata column will be \code{"mean"} if no track was provided, or
\code{"mean_<track>"} if a track was provided.}
}
\value{
New version of \code{granges} with region mean added as a metadata
attribute for each genomic range.
}
\description{
For a set of genomic ranges in a \code{GRanges} object, adds the mean value over
the region as a metadata attribute for each genomic range, returning a new
\code{GRanges} object with updated metadata.
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
granges <- GenomicRanges::GRanges(
             seqnames = c('chr1'), 
             ranges = IRanges::IRanges(1000:2000), 
             strand = c('+'))
granges_with_mean <- update_with_mean(granges, d4source)

}
