% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{update_with_query_results}
\alias{update_with_query_results}
\title{Add D4 query results to a \code{GRanges} object}
\usage{
update_with_query_results(granges, d4source, ..., track = NA)
}
\arguments{
\item{granges}{\code{GRanges} object containing genomic ranges of interest}

\item{d4source}{\code{D4Source} object}

\item{...}{Not used for values, forces later arguments to bind by name}

\item{track}{Track name. If omitted or NA, queries the first track.}
}
\value{
New version of \code{granges} with query results added as the
\code{'query_results'} metadata attribute for each genomic range. For ranges on
the minus strand, the vector of position scores reads in the minus strand
direction. If a track is specified, the track name will be included in the
metadata column name.
}
\description{
For a set of genomic ranges in a \code{GRanges} object, extracts each region from
the D4 file and adds the query result as a metadata attribute for each
genomic range, returning a new \code{GRanges} object with updated metadata.
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
granges <- GenomicRanges::GRanges(
             seqnames = c('chr1'), 
             ranges = IRanges::IRanges(1000:2000), 
             strand = c('+'))
granges_with_query <- update_with_query_results(granges, d4source)

}
