% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{D4Source}
\alias{D4Source}
\title{Establish a connection to a D4 file}
\usage{
D4Source(path)
}
\arguments{
\item{path}{File path}
}
\value{
Object of class \code{D4Source}
}
\description{
Establish a connection to a D4 file
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
methods(class = 'D4Source')

}
