% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{get_chroms}
\alias{get_chroms}
\title{Generic function for the chromosomes in a data object}
\usage{
get_chroms(x)
}
\arguments{
\item{x}{Input object}
}
\value{
List-like container of chromosomes
}
\description{
Generic function for the chromosomes in a data object
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
get_chroms(d4source)

}
