% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{get_source}
\alias{get_source}
\title{Generic function for the source of an object}
\usage{
get_source(x)
}
\arguments{
\item{x}{Input object}
}
\value{
The source
}
\description{
Generic function for the source of an object
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
get_source(d4source)

}
