% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{get_tracks}
\alias{get_tracks}
\title{Generic function for the tracks in a data object}
\usage{
get_tracks(x)
}
\arguments{
\item{x}{Input object}
}
\value{
List-like container of track names
}
\description{
Generic function for the tracks in a data object
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
get_tracks(d4source)

}
