% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{mean.D4Source}
\alias{mean.D4Source}
\title{Returns the mean data value over the given region}
\usage{
\method{mean}{D4Source}(x, chr, start, end, ..., track = NA)
}
\arguments{
\item{x}{\code{D4Source} object}

\item{chr}{Chromosome/contig name}

\item{start}{0-based start position}

\item{end}{0-based exclusive end position}

\item{...}{Not used for values, forces later arguments to bind by name}

\item{track}{Track name. If omitted or NA, queries the first track.}
}
\value{
Mean data value
}
\description{
Returns the mean data value over the given region
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
mean(d4source, 'chr1', 1000, 2000)

}
