% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{median.D4Source}
\alias{median.D4Source}
\title{Returns the median data value over the given region}
\usage{
median.D4Source(d4source, chr, start, end, ..., track = NA)
}
\arguments{
\item{d4source}{\code{D4Source} object}

\item{chr}{Chromosome/contig name}

\item{start}{0-based start position}

\item{end}{0-based exclusive end position}

\item{track}{Track name. If omitted or NA, queries the first track.}
}
\value{
Median data value
}
\description{
Returns the median data value over the given region
}
\examples{
d4source <- D4Source(system.file("extdata", "example.d4", package="rd4"))
median(d4source, "chr1", 1000, 2000)

}
