% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{percentile}
\alias{percentile}
\title{Generic function to calculate percentile over a genomic interval}
\usage{
percentile(x, chr, start, end, percentile, ..., track)
}
\arguments{
\item{x}{Input object}

\item{chr}{Chromosome/contig name}

\item{start}{0-based start position}

\item{end}{0-based exclusive end position}

\item{percentile}{Percentile between 0 and 100. Example: percentile 0 is
the minimum data value, 50 is the median and 100 is the maximum.}

\item{...}{Not used for values, forces later arguments to bind by name}

\item{track}{Track name}
}
\value{
Data value at the given percentile
}
\description{
Generic function to calculate percentile over a genomic interval
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
percentile(d4source, 'chr1', 1000, 2000, 50)

}
