% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{print.D4Source}
\alias{print.D4Source}
\title{Prints information about a \code{D4Source} object}
\usage{
\method{print}{D4Source}(d4source)
}
\arguments{
\item{d4source}{\code{D4Source} object}
}
\value{
(Invisibly) the string that was printed
}
\description{
Prints information about a \code{D4Source} object
}
