% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Query}
\alias{Query}
\alias{$.Query}
\title{A context object that holds onto the original query parameters}
\format{
An object of class \code{environment} of length 3.
}
\usage{
Query
}
\description{
A context object that holds onto the original query parameters
}
\keyword{datasets}
