% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{resample}
\alias{resample}
\title{Bin across the given region, populating each bin with a summary of the
original data values in the bin region}
\usage{
resample(
  d4source,
  chr,
  start,
  end,
  method,
  ...,
  bin_size = 1000,
  allow_bin_size_adjustment = TRUE,
  track = NA
)
}
\arguments{
\item{d4source}{\code{D4Source} object}

\item{chr}{Chromosome/contig name}

\item{start}{0-based start position}

\item{end}{0-based exclusive end position}

\item{method}{Summary statistic for the data in each bin; one of 'mean' or
'median'}

\item{...}{Not used for values, forces later arguments to bind by name}

\item{bin_size}{Bin size}

\item{allow_bin_size_adjustment}{Only used for remote files: optimize bin
size for reading over a network}

\item{track}{Track name. If omitted or NA, queries the first track.}
}
\value{
New query result of resampled data as a list
}
\description{
Bin across the given region, populating each bin with a summary of the
original data values in the bin region
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
resample(d4source, 'chr1', 0, 50000000, 'mean')

}
