% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4source.R
\name{update_mean}
\alias{update_mean}
\title{Generic function to update mean metadata in a GRanges object}
\usage{
update_mean(x, granges, ..., track)
}
\arguments{
\item{x}{Input object}

\item{granges}{GRanges object}

\item{...}{Not used for values, forces later arguments to bind by name}

\item{track}{Track name}
}
\value{
New version of \code{granges} with metadata updated
}
\description{
Generic function to update mean metadata in a GRanges object
}
\examples{
d4source <- D4Source(system.file('extdata', 'example.d4', package = 'rd4'))
granges <- GRanges(seqnames = c('chr1'), ranges = IRanges(1000:2000), 
                   strand = c('+'))
granges_with_mean <- update_mean(d4source, granges)

}
